// src/lib/api/generated/main/models/fileMetadataResponse.ts
/**
 * Generated by orval v7.5.0 🍺
 * Do not edit manually.
 * Devlens Admin API 문서
 * Devlens Admin API 문서
 * OpenAPI spec version: 1.0
 */

export interface FileMetadataResponse {
  id?: number;
  category?: string;
  referenceId?: number;
  displayTitle?: string;
  title?: string;
  contentType?: string;
  format?: string;
  size?: number;
  path?: string;
  createdBy?: number;
  createdAt?: string;
}